+++ src/constants/flags.ts
@@ -1,4 +1,4 @@
-export const flags: { [key: string]: { flag: string; value?: string }[] } = {
-  gpuRasterization: [{ flag: "enable-gpu-rasterization", value: undefined }],
-  disableHardwareMediaKeys: [{ flag: "disable-features", value: "HardwareMediaKeyHandling" }],
+export const flags: { [key: string]: { flag: string; value?: string } } = {
+  gpuRasterization: { flag: "enable-gpu-rasterization", value: undefined },
+  disableHardwareMediaKeys: { flag: "disable-features", value: "HardwareMediaKeyHandling" },
 };